network Error view IMPL started
