+++ client/src/components/pages/single/single.scss
@@ -1,3 +1,20 @@
 .ui-single {
     display: flex;
+    min-height: calc(100vh - 50px);
+    width: calc( 100vw - 60px );
+    background-color: rgb(247, 245, 245);
+    gap: 10px;
+    padding: 20px 30px;
+
+    &__widget {
+        background-color: #fff;
+
+        &__singlePost {
+            flex: 3;
+        }
+
+        &__sidebar {
+            flex: 1;
+        }
+    }
 }
