+++ client/src/generic/networkError/networkError.scss
@@ -1,0 +1,7 @@
+.ui-network-error {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
